backend/requirements.txt
asgiref==3.9.1
Django==4.2.7
django-cors-headers==4.7.0
djangorestframework==3.16.0
pillow==11.3.0
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.3
